app/requirements.txt
asgiref==3.6.0
Django==4.1.4
psycopg2-binary==2.9.5
sqlparse==0.4.3
gunicorn==20.1.0
celery==5.2.7
redis==4.4.0
flower==1.2.0
boto3==1.26.59
django-storages==1.13.2
djangorestframework==3.12.4
requests==2.26.0
django-redis==5.0.0
pymemcache==3.5.0
django-lifecycle==0.9.1
bpython==0.24
django-extensions==3.2.1
ipython==8.12.0
django-cors-headers==3.14.0
graphene-django==3.0.0

# Elasticsearch
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
django-elasticsearch-dsl==7.3

# Testing
pytest==7.2.0
pytest-django==4.5.2
Faker==8.8.0
django-debug-toolbar==3.2.1
black==21.6b0